backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==21.2.0
python-multipart==0.0.20
pydantic[email]==2.10.5
python-dateutil==2.9.0
lxml==5.3.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
aiofiles==24.1.0

